Keep the stale floor for seq-less replay tails until their ack
A compatibility host can answer a replay with a raw byte tail carrying no
sequence; re-basing the stale floor on that acceptance defeated the ack
path that restores the floor as the baseline for exactly this case,
leaving the surface baseline-less after the barrier cleared. Only a
sequence-carrying acceptance re-bases the floor. Byte-channel floor
tests move to their own file.
